remove vue-router
