--- src/views/index.ts
@@ -1,13 +1,0 @@
-import type { RouteRecordRaw } from 'vue-router';
-
-import Home from './HomeView.vue';
-
-const routes: RouteRecordRaw[] = [
-  {
-    path: '/',
-    name: 'home',
-    component: Home,
-  },
-];
-
-export { routes };
